hikey/hikey960: Enable drm_hwcomposer for 4.14+ kernels
Modifications needed to enable the drm_hwcomposer in
the hikey build.

This enables drm_hwcomposer for 4.14+ on HiKey and HiKey960

This requires updated drm_hwc source and changes to the kernel
source and 4.14 kernel prebuilt to be updated (and merged
together with this change).

So DO NOT MERGE until pre-built kernel which incldues
the following commits is submitted:
https://android-review.googlesource.com/c/kernel/hikey-linaro/+/743923

As well as the following update to drm_hwcomposer:
https://android-review.googlesource.com/c/platform/external/drm_hwcomposer/+/743902

Bug: 62066689
Change-Id: Ifbe76eed77672d06ceb3bce16de4b4d8cac56f17
Signed-off-by: John Stultz <john.stultz@linaro.org>