hikey: Fix power-hal logcat noise
The way powerhal modules are initialized maybe changed
or was always broken, but as it is, some of the hikey power
module entries were not being initialized properly.

This was causing lots of noise in the logs, and potentially
other problematic behavior.

This changes the initial allocation so we allocate the whole
structure and initialize it properly.

Change-Id: I7d7858c2890d03758b302b95f18bd1c35b6ef85b
Signed-off-by: John Stultz <john.stultz@linaro.org>