hikey960: Add version check on the hisi vdec package
This adds a version check on the hisi vdec package and provides
a URL to the current expected package.

This did get a bit hacky as we can't check values set in the
vendor pkg makefile if we use inherit-product. So we have to
include it directly. I think we could do it better, but we would
have to re-spin the published vendor binary, so lets hack around
it this time and do better with the next package update.

Change-Id: Ic6832850e9b4d03b84358060810aa17e23e5b8ae
Signed-off-by: John Stultz <john.stultz@linaro.org>