hikey{,960}: Enable system-as-root and verity.
Add a build option TARGET_ANDROID_VERITY which can be set to anything
but the empty string to enable dm-android-verity. This option is
introduced because the UEFI bootloader for hikey and hikey960 has a bug
where it incorrectly rejects boot.img files with no ramdisk (see
https://bugs.96boards.org/show_bug.cgi?id=757).

This change *will* work on a hikey960 with the non-UEFI bootloader
installed.

Test: local
Bug: 72722987
Bug: 111829702
Change-Id: I75d0941866d9196e78e7c2943cdf05d35beb1879
Signed-off-by: Alistair Strachan <astrachan@google.com>