hikey960: Fix image sizes to match ptable
The partition table blob defines partition sizes which do not match the
size of the images created by the Android build system. This tickles a
bug in the libfec userspace code, which is fixed in system/extras
Ia6cdbc0f35a98fdfa1fb4677951ae5c829cabb0f, but we may as well stop
wasting flash space.

Test: local
Bug: 72722987
Change-Id: Ia971c0c3ae5b3e798b107a8c6d571cb67d8ec0b9
Signed-off-by: Alistair Strachan <astrachan@google.com>