device-common.mk: set PRODUCT_SHIPPING_API_LEVEL
to make CTS android.os.cts.BuildTest#testSdkInt pass

Here we set it to 25 first, since set to 26 or later needs
to support gatekeeper function, which makes the build
failed to boot to homescreen with the default aosp implementation.

Will set it to 28 or the latest api number when we have all the
necessary features supported.

Test: cts-tradefed run commandAndExit cts -m CtsOsTestCases -t android.os.cts.BuildTest#testSdkInt

Change-Id: Ibdebe816e24674599488bc36d6fe9ade2d152613
Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>