hikey/hikey960: Remove use of FIQ0 console
Since the FIQ_debugger is now deprecated in android-4.19,
and its utility on hikey/hikey960 has always been minimal
given there isn't a real FIQ (just an maskable IRQ), lets
just remove the use of the FIQ0 console.

We still have to keep it for 4.4/4.9/4.14 hikey kernels,
as removing it breaks console input. We will need to
land kernel changes and update userspace.

But for hikey960 removing it is ok.

Change-Id: I85ac12660ab0b5d38d64486ffef4ae0cebbe5e8f
Signed-off-by: John Stultz <john.stultz@linaro.org>