HiKey960: Set driver_match_existing_only to 0 to enable gadget mode to work on usb-c hotplug
With the upstream targetted patches, if the system boots up and the
usb-c cable is unplugged, when it is plugged in the system will not
fully switch to gadget mode, causing adb to fail. This is due to
the gadget enumerating on a different device number then what is
expected.

Thus this patch sets driver_match_existing_only to 0 to allow
the new number to be used.

Also updates init.recovery.common.rc file

EDIT: Changes in init.recovery.common.rc REMOVED

Suggested by: Chen Yu <chenyu56@huawei.com>
Signed-off-by: John Stultz <john.stultz@linaro.org>

Change-Id: I217653edc013f888feb8d97220e466903834607b
Signed-off-by: Victor Chong <victor.chong@linaro.org>