hikey: gralloc: Add element to allow upstream drm_hwc to build w/ HiKey
Upstream drm_hwc merged some changes to support AFBC which is
only supported on HiKey960. In doing so they utilize a hikey960
gralloc private handle field that doesn't exist on HiKey.

Thus this patch adds the internal_format field as a union
element on format. This is a union element so we don't change
the structure size and since internal_format isn't used in
drm_hwc when building with w/ HiKey there is no effect on the
values it aliases.

Change-Id: Ie73706d8f3acfaf3bfa5c187c0365d505da044d2
Signed-off-by: John Stultz <john.stultz@linaro.org>