gralloc: Clean up previous ion api rework for hikey gralloc
Some of the work done for the ION 4.12+ ABI change
could be done in a cleaner fashion.

This tries to consolidate the logic so we have less
conditional switching depending on if we're using
legacy code or not, and also utilizes ion_alloc_fd()
to simplify the older ion_alloc usage.

This is useful as groundwork for switching to DMABUF
heaps

Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: Ie917406646a5eb802800d3f0eaff95fa591af0db