fstab.hikey960: make it possible for vendor partition to be remounted
as the "adb remount" checks the /vendor/etc/fstab.hikey960
file to find if the partition is a known partition,
so we need to add vendor here to make it work with adb remount.

Otherwise will get following error:
$ adb remount vendor
Unknown partition vendor, skipping
No partitions to remount
remount failed
$

Or it's not remounted to be writable after adb remount command.

Test: adb remount, and vendor partition is writable.

Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>
Change-Id: I4a6e0ddc31cd62bcf9b7c42083656c7dbaed5009