Remove uefi related artifacts
The uefi source repos are being removed from the manifest, therefore
we must remove binaries derived from these sources.

This removes the hikey bootloader completely, but hikey is not really
available any more and the old installer can be fetched from elsewhere.

For hikey960, uefi was never stable, and the hisi blobs must be used
instead. Remove the uefi bootloader binaries for hikey960.

Change-Id: I77fa2ef9017a770777d15b5d28880619cb953b6c